refactor(trusty-common): move the drawer slot-index helpers out of write_ops (#4810)
write_ops.rs sat at 495 of its 500 permitted SLOC, leaving no room for the
#4810 triple-key work. The drawer half — stored_fact_key,
release_slot_if_owned_by, batch_upsert_drawer, batch_delete_drawer — shares
nothing with the triple write path but a transaction, so it moves verbatim to
a sibling kg_redb/drawer_ops.rs. write_ops drops to 434 SLOC; drawer_ops is
68. No behavior change: the four functions are byte-identical apart from the
import block.

Refs #4810
